+++ repository/src/main/java/com/netcracker/repositories/impl/BookmarkRepositoryImpl.java
@@ -40,7 +40,7 @@ public class BookmarkRepositoryImpl implements IRepository<Bookmark> {
     }
 
     @Override
-    public Bookmark getById(int id) {
+    public Bookmark getById(long id) {
         return null;
     }
