+++ repository/src/main/java/com/netcracker/repositories/impl/SocialNetworkRepositoryImpl.java
@@ -26,7 +26,7 @@ public class SocialNetworkRepositoryImpl implements IRepository<SocialNetwork> {
     }
 
     @Override
-    public SocialNetwork getById(int id) {
+    public SocialNetwork getById(long id) {
         return null;
     }
